Refactor region description analyzers
