+++ Reihitsu.Analyzer/Base/RegionDescriptionAnalyzerBase{TAnalyzer}.cs
@@ -1,0 +1,114 @@
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Diagnostics;
+
+using Reihitsu.Analyzer.Enumerations;
+
+namespace Reihitsu.Analyzer.Base;
+
+/// <summary>
+/// Base class for analyzers that validate region descriptions
+/// </summary>
+/// <typeparam name="TAnalyzer">Type of the analyzer</typeparam>
+public abstract class RegionDescriptionAnalyzerBase<TAnalyzer> : DiagnosticAnalyzerBase<TAnalyzer>
+    where TAnalyzer : DiagnosticAnalyzer
+{
+    #region Constructor
+
+    /// <summary>
+    /// Constructor
+    /// </summary>
+    /// <param name="diagnosticId">Diagnostic ID</param>
+    /// <param name="category">Category</param>
+    /// <param name="titleResourceName">Resource name of the title</param>
+    /// <param name="messageFormatResourceName">Resource name of the message format</param>
+    private protected RegionDescriptionAnalyzerBase(string diagnosticId, DiagnosticCategory category, string titleResourceName, string messageFormatResourceName)
+        : base(diagnosticId, category, titleResourceName, messageFormatResourceName)
+    {
+    }
+
+    #endregion // Constructor
+
+    #region Methods
+
+    /// <summary>
+    /// Gets the description of an endregion directive
+    /// </summary>
+    /// <param name="directive">Directive</param>
+    /// <returns>Description text</returns>
+    private static string GetEndRegionDescription(EndRegionDirectiveTriviaSyntax directive)
+    {
+        var description = (directive.EndRegionKeyword.TrailingTrivia.ToFullString()
+                           + directive.EndOfDirectiveToken.LeadingTrivia.ToFullString()).Trim();
+
+        if (description.StartsWith("//", StringComparison.Ordinal))
+        {
+            description = description.Substring(2).TrimStart();
+        }
+
+        return description;
+    }
+
+    /// <summary>
+    /// Gets the description of a region directive
+    /// </summary>
+    /// <param name="directive">Directive</param>
+    /// <returns>Description text</returns>
+    private static string GetRegionDescription(RegionDirectiveTriviaSyntax directive)
+    {
+        var messageTrivia = directive.EndOfDirectiveToken.LeadingTrivia.FirstOrDefault(static trivia => trivia.IsKind(SyntaxKind.PreprocessingMessageTrivia));
+
+        return messageTrivia == default
+                   ? string.Empty
+                   : messageTrivia.ToString().Trim();
+    }
+
+    /// <summary>
+    /// Determines whether the specified description violates the rule
+    /// </summary>
+    /// <param name="description">Description to inspect</param>
+    /// <returns><see langword="true"/> if the description violates the rule</returns>
+    protected abstract bool IsInvalidDescription(string description);
+
+    /// <summary>
+    /// Analyzes endregion directives
+    /// </summary>
+    /// <param name="context">Context</param>
+    private void OnEndRegion(SyntaxNodeAnalysisContext context)
+    {
+        if (context.Node is EndRegionDirectiveTriviaSyntax node
+            && IsInvalidDescription(GetEndRegionDescription(node)))
+        {
+            context.ReportDiagnostic(CreateDiagnostic(node.GetLocation()));
+        }
+    }
+
+    /// <summary>
+    /// Analyzes region directives
+    /// </summary>
+    /// <param name="context">Context</param>
+    private void OnRegion(SyntaxNodeAnalysisContext context)
+    {
+        if (context.Node is RegionDirectiveTriviaSyntax node
+            && IsInvalidDescription(GetRegionDescription(node)))
+        {
+            context.ReportDiagnostic(CreateDiagnostic(node.GetLocation()));
+        }
+    }
+
+    #endregion // Methods
+
+    #region DiagnosticAnalyzer
+
+    /// <inheritdoc/>
+    public override void Initialize(AnalysisContext context)
+    {
+        base.Initialize(context);
+
+        context.RegisterSyntaxNodeAction(OnRegion, SyntaxKind.RegionDirectiveTrivia);
+        context.RegisterSyntaxNodeAction(OnEndRegion, SyntaxKind.EndRegionDirectiveTrivia);
+    }
+
+    #endregion // DiagnosticAnalyzer
+}
